feat: 자동 로그인시 바로 마이페이지로 이동
