+++ frontend/src/pages/LoginOptionPage.js
@@ -7,10 +7,33 @@ import { useSelector } from 'react-redux'
 import styles from '@/styles'
 import Back from '@/components/Back'
 import { loginOption } from '@/constants/language'
+import { removeJWT, setJWT } from '@/util'
+import asyncStorage from '@react-native-async-storage/async-storage/src/AsyncStorage'
+import { reAuthentication } from '@/api/api'
 
 const LoginOption = () => {
   const navigation = useNavigation()
   const language = useSelector((state) => state.languageOption)
+
+  const newToken = (token) => {
+    reAuthentication(token)
+      .then((result) => {
+        if (result.status === 200) removeJWT().then(setJWT(result.data).then(() => navigation.navigate('Mypage')))
+      })
+      .catch(() => {
+        removeJWT()
+        navigation.navigate('Login')
+      })
+  }
+
+  const login = () => {
+    asyncStorage
+      .getItem('JWT')
+      .then((value) => {
+        newToken(value)
+      })
+      .catch(() => navigation.navigate('Login'))
+  }
 
   return (
     <View style={styles.container}>
@@ -25,7 +48,7 @@ const LoginOption = () => {
           {Logo()}
         </View>
         <View style={LoginOptionStyles.buttons}>
-          <TouchableOpacity style={styles.longBox} onPress={() => navigation.navigate('Login')}>
+          <TouchableOpacity style={styles.longBox} onPress={login}>
             <Text style={styles.boxInnerText}>{loginOption[language].login}</Text>
           </TouchableOpacity>
           <TouchableOpacity style={styles.longBox} onPress={() => navigation.navigate('Signup')}>
